Move Property validators
